kopierat det nya projektet

+++ src/main/java/com/fredrikkodar/blogwebservice/service/UserService.java
@@ -3,13 +3,21 @@ package com.fredrikkodar.blogwebservice.service;
 import com.fredrikkodar.blogwebservice.models.User;
 import com.fredrikkodar.blogwebservice.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
-public class UserService {
+public class UserService implements UserDetailsService {
+
+    @Autowired
+    private PasswordEncoder encoder;
 
     @Autowired
     private UserRepository userRepository;
@@ -22,16 +30,23 @@ public class UserService {
         return userRepository.findById(id).orElse(null);
     }
 
-    public User registerUser(User user) {
-        return userRepository.save(user);
-    }
-
-    public User logInUser(String username, String password) {
-        Optional<User> user = userRepository.findByUsername(username);
-        return user.filter(value -> value.getPassword().equals(password)).orElse(null);
-    }
-
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        }
+        else {
+            throw new IllegalArgumentException("User with id " + id + " does not exist");
+        }
+    }
+
+    public User getCurrentUser() {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        Optional<User> userOptional = userRepository.findByUsername(username);
+        return userOptional.orElseThrow(() -> new UsernameNotFoundException("User not found with username: " + username));
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException { // Hämtar användare från databasen om den finns
+        return userRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("User not found with username: " + username));
     }
 }
